Import Project FeederWatch Data

This function reads all .csv files downloaded from the Project FeederWatch website, either from the default "data-raw/" folder or from a user-specified folder. Optionally, it can apply filters like region, species, year, etc.

pfw_import(folder = "data-raw/", filter = FALSE, ...)

Arguments

folder

The folder where Project FeederWatch data is stored. Default is "data-raw/".

filter

Logical. If TRUE, applies filters using pfw_filter(). Default is FALSE.

...

Additional arguments passed to pfw_filter() for filtering (e.g., region, species, year).

Value

A combined and optionally filtered data.table containing all Project FeederWatch data.

Examples

if (FALSE) { # \dontrun{
# Import all downloaded data from the default folder ("data-raw")
data <- pfw_import()

# Import and filter for Washington checklists from 2023
data_filtered <- pfw_import(filter = TRUE, region = "Washington", year = 2023)
} # }